+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.1.24
+
+### Features
+
+- added seeding prompts to assist with seeding django models.
+
+### Fixes
+
+- A lot of refactoring for the seeding app.
+
 ## v0.1.23
 
 ### Changes